agent/pyproject.toml
[project]
name = "amis-ai-agent"
version = "0.1.0"
description = "amis-ai 智能体服务 — AI 辅助生成 amis JSON 配置"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "asyncpg>=0.30.0",
    "sse-starlette>=2.2.0",
    "python-dotenv>=1.0.0",
]
